refactor variant logic out of ingestion listener and into variant_utils
